# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Sentinel AWS - Cloud Security Auditing Tool API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface RegisterRequest {
  name: string;
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface UpdateProfileRequest {
  name?: string;
  email?: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface CreateAccountRequest {
  name: string;
  accessKeyId: string;
  secretAccessKey: string;
  region: string;
}

export interface UpdateAccountRequest {
  name?: string;
  accessKeyId?: string;
  secretAccessKey?: string;
  region?: string;
}

export interface AwsAccount {
  id: number;
  name: string;
  accessKeyId: string;
  region: string;
  /** @nullable */
  lastScanAt?: string | null;
  /** @nullable */
  lastScore?: number | null;
  createdAt: string;
}

export interface ConnectionTestResult {
  success: boolean;
  /** @nullable */
  accountId?: string | null;
  message: string;
}

export type StartScanRequestModulesItem = typeof StartScanRequestModulesItem[keyof typeof StartScanRequestModulesItem];


export const StartScanRequestModulesItem = {
  iam: 'iam',
  s3: 's3',
  ec2: 'ec2',
  security_groups: 'security_groups',
  cloudtrail: 'cloudtrail',
} as const;

export interface StartScanRequest {
  accountId: number;
  modules?: StartScanRequestModulesItem[];
}

export type ScanStatus = typeof ScanStatus[keyof typeof ScanStatus];


export const ScanStatus = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface Scan {
  id: number;
  accountId: number;
  accountName: string;
  status: ScanStatus;
  progress: number;
  /** @nullable */
  currentModule?: string | null;
  /** @nullable */
  overallScore?: number | null;
  /** @nullable */
  iamScore?: number | null;
  /** @nullable */
  s3Score?: number | null;
  /** @nullable */
  ec2Score?: number | null;
  /** @nullable */
  networkingScore?: number | null;
  /** @nullable */
  cloudtrailScore?: number | null;
  criticalCount: number;
  highCount: number;
  mediumCount: number;
  lowCount: number;
  startedAt: string;
  /** @nullable */
  completedAt?: string | null;
  /** @nullable */
  errorMessage?: string | null;
}

export type FindingService = typeof FindingService[keyof typeof FindingService];


export const FindingService = {
  iam: 'iam',
  s3: 's3',
  ec2: 'ec2',
  security_groups: 'security_groups',
  cloudtrail: 'cloudtrail',
} as const;

export type FindingSeverity = typeof FindingSeverity[keyof typeof FindingSeverity];


export const FindingSeverity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
  informational: 'informational',
} as const;

export interface Finding {
  id: number;
  scanId: number;
  service: FindingService;
  checkId: string;
  title: string;
  description: string;
  severity: FindingSeverity;
  affectedResource: string;
  recommendation: string;
  /** @nullable */
  metadataJson?: string | null;
}

export type ScanDetail = Scan & {
  findings: Finding[];
};

export interface ScanListResponse {
  data: Scan[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface ScanComparison {
  current: Scan;
  previous: Scan;
  scoreDelta: number;
  newFindings: Finding[];
  resolvedFindings: Finding[];
  persistingFindings: Finding[];
}

export interface FindingListResponse {
  data: Finding[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface ServiceScore {
  service: string;
  score: number;
  criticalCount: number;
  highCount: number;
  mediumCount: number;
  lowCount: number;
}

export interface ScoreHistoryPoint {
  date: string;
  score: number;
}

export interface DashboardSummary {
  /** @nullable */
  overallScore?: number | null;
  /** @nullable */
  lastScanAt?: string | null;
  totalAccounts: number;
  totalScans: number;
  criticalCount: number;
  highCount: number;
  mediumCount: number;
  lowCount: number;
  serviceScores: ServiceScore[];
  recentScans: Scan[];
  recentFindings: Finding[];
  scoreHistory: ScoreHistoryPoint[];
}

export interface ReportSummary {
  scan: Scan;
  criticalFindings: Finding[];
  highFindings: Finding[];
  mediumFindings: Finding[];
  lowFindings: Finding[];
}

export type ListScansParams = {
accountId?: number;
page?: number;
limit?: number;
};

export type ListFindingsParams = {
scanId?: number;
accountId?: number;
severity?: ListFindingsSeverity;
service?: ListFindingsService;
search?: string;
page?: number;
limit?: number;
};

export type ListFindingsSeverity = typeof ListFindingsSeverity[keyof typeof ListFindingsSeverity];


export const ListFindingsSeverity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
  informational: 'informational',
} as const;

export type ListFindingsService = typeof ListFindingsService[keyof typeof ListFindingsService];


export const ListFindingsService = {
  iam: 'iam',
  s3: 's3',
  ec2: 'ec2',
  security_groups: 'security_groups',
  cloudtrail: 'cloudtrail',
} as const;

export type GetDashboardSummaryParams = {
accountId?: number;
};

export type DownloadCsvReportParams = {
scanId?: number;
accountId?: number;
};

export type GetReportSummaryParams = {
scanId?: number;
accountId?: number;
};