# lib/api-zod/src/generated/types/awsAccount.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Sentinel AWS - Cloud Security Auditing Tool API
 * OpenAPI spec version: 0.1.0
 */

export interface AwsAccount {
  id: number;
  name: string;
  accessKeyId: string;
  region: string;
  /** @nullable */
  lastScanAt?: string | null;
  /** @nullable */
  lastScore?: number | null;
  createdAt: string;
}